优化代码 新增shader和resource的favorite和note接口

+++ block-snap-server/block-snap-common/src/main/java/com/github/doyxs/blocksnap/common/api/Result.java
@@ -6,14 +6,14 @@ import lombok.Data;
 @Data
 public class Result<T> {
     
-    private long code;    // 状态码
+    private int code;    // HTTP 风格状态码（200/401/403/404/500），用 int 避免 JSON 序列化为 long 时前端精度问题
     private String message; // 提示信息
     private T data;       // 真正返回给前端的数据
 
     // 私有化构造方法，强制通过后面的静态方法来创建对象
     protected Result() {}
 
-    protected Result(long code, String message, T data) {
+    protected Result(int code, String message, T data) {
         this.code = code;
         this.message = message;
         this.data = data;
